Ram Refactoring for Ram Aggregates; improve condition hoisting; refactored RAM transforms; issues with nullaries in AstTranslator

+++ src/RamConditionLevel.h
@@ -36,7 +36,7 @@ public:
     void run(const RamTranslationUnit& translationUnit) override;
 
     /** Get level */
-    size_t getLevel(const RamCondition* condition) const;
+    int getLevel(const RamCondition* condition) const;
 
 protected:
     RamExpressionLevelAnalysis* rvla;
